requirements.txt
yfinance
riskfolio-lib
pandas
matplotlib
